+++ engine/mindol/tools/__init__.py
@@ -1,1 +1,0 @@
-"""mindol.tools - 工具集"""