+++ engine/shalou/tools/__init__.py
@@ -1,0 +1,1 @@
+"""shalou.tools - 工具集"""